Gemma 4 12B default + UI compaction, compact GPU panel, admin tooling
Model / infra:
- Default to google/gemma-4-12B (base, multimodal) loaded 8-bit (~13GB, fits 16GB);
  transformers>=5.10 in an isolated project .venv (arena-env left untouched).
- testbed.py: CLI to load any model + check it predicts before wiring into the server.
- Per-keystroke coalescing: /ws reader/worker skips stale prefixes under fast typing.

Admin / ops:
- Admin toggle to enable/disable the live conversation viewer (server-enforced).
- Reboot button (in-place execv restart; reloads code + model).
- Default token cap 1024.
- Cache-Control: no-cache on HTML/JS/CSS (+ ?v= busting) so the Cloudflare tunnel
  stops serving stale assets.

UI:
- Compacted to fit 1080p; live-feed link under the editor; accuracy + token count
  boxed in the header top-right; terminals side-by-side.
- GPU monitor rebuilt as a compact JSON-fed panel: one info line, horizontal ASCII
  bars for MEM and current Util%, and a fixed 32-bar util sparkline over time.

Docs: README + project skill file (incl. branch-first git workflow).

+++ predictors.py
@@ -107,7 +107,8 @@ class HFPredictor(BasePredictor):
 
             quant_config = BitsAndBytesConfig(load_in_8bit=True)
 
-        common = dict(torch_dtype=dtype, low_cpu_mem_usage=True)
+        # transformers 5.x renamed `torch_dtype` -> `dtype` (the old name warns).
+        common = dict(dtype=dtype, low_cpu_mem_usage=True)
         n_gpus = torch.cuda.device_count() if self.device == "cuda" else 0
 
         if quant_config is not None:
